+++ lib/src/dependency.g.dart
@@ -12,6 +12,8 @@ part of 'dependency.dart';
 
 SdkDependency _$SdkDependencyFromJson(Map json) {
   return $checkedNew('SdkDependency', json, () {
+    $checkKeys(json,
+        requiredKeys: const ['sdk'], disallowNullValues: const ['sdk']);
     var val = new SdkDependency(
         $checkedConvert(json, 'sdk', (v) => v as String),
         version: $checkedConvert(json, 'version',
@@ -26,7 +28,7 @@ GitDependency _$GitDependencyFromJson(Map json) {
         requiredKeys: const ['url'], disallowNullValues: const ['url']);
     var val = new GitDependency(
         $checkedConvert(
-            json, 'url', (v) => v == null ? null : _parseUri(v as String)),
+            json, 'url', (v) => v == null ? null : parseGitUri(v as String)),
         $checkedConvert(json, 'ref', (v) => v as String),
         $checkedConvert(json, 'path', (v) => v as String));
     return val;
@@ -56,7 +58,7 @@ HostedDetails _$HostedDetailsFromJson(Map json) {
     var val = new HostedDetails(
         $checkedConvert(json, 'name', (v) => v as String),
         $checkedConvert(
-            json, 'url', (v) => v == null ? null : _parseUri(v as String)));
+            json, 'url', (v) => v == null ? null : parseGitUri(v as String)));
     return val;
   });
 }