+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.2
+
+- Allow superfluous `version` keys with `git` and `path` dependencies.
+- Improve errors when unsupported keys are provided in dependencies.
+- Provide better errors with invalid `sdk` dependency values.
+- Support "scp-like syntax" for Git SSH URIs in the form
+  `[user@]host.xz:path/to/repo.git/`.
+
 ## 0.1.1
 
 - Fixed name collision with error type in latest `package:json_annotation`.